Refactoring shared code between pond online and offline.
