+++ .gitignore
@@ -4,6 +4,7 @@
 appengine/generated/
 appengine/*/generated/
 appengine/*/*/generated/
+appengine/*/*/*/generated/
 appengine/third-party/
 json/keys.json
 third-party-downloads/*
