+++ appengine/about.html
@@ -92,7 +92,7 @@ between blocks and actual JavaScript in a text editor.</p>
 
   <tr>
     <td>
-      <img src="index/pond-duck.png" height=100 width=100>
+      <img src="index/pond-duck-basic.png" height=100 width=100>
     </td>
     <td>
 <p>Pond is an open-ended contest to program the smartest duck.
