+++ appengine/common/boot.js
@@ -30,7 +30,10 @@ if (location.host == 'blockly-games.appspot.com') {
 (function() {
   // Application path.
   var appName = location.pathname.match(/\/([-\w]+)(\.html)?$/);
-  appName = appName ? appName[1].replace('-', '/') : 'index';
+  console.log(appName);
+  appName = appName ? appName[1].replace(/-/g, '/') : 'index';
+  console.log(appName);
+
 
   // Supported languages (consistent across all apps).
   window['BlocklyGamesLanguages'] = [
