+++ appengine/pond-duck-online.html
@@ -4,7 +4,7 @@
   <meta charset="utf-8">
   <meta name="google" value="notranslate">
   <meta name="viewport" content="target-densitydpi=device-dpi, width=device-width, initial-scale=1.0, user-scalable=no">
-  <title>Blockly Games : Pond</title>
+  <title>Blockly Games : Pond Online</title>
   <link rel="stylesheet" href="common/common.css">
   <link rel="stylesheet" href="pond/style.css">
   <link rel="stylesheet" href="pond/duck/style.css">
