+++ appengine/pond/duck/online/template.soy
@@ -1,4 +1,4 @@
-{namespace Pond.Online.soy}
+{namespace Pond.Duck.Online.soy}
 
 /**
  * This is a Closure Template.
@@ -32,7 +32,7 @@
   <h1>
     {call BlocklyGames.soy.titleSpan}
       {param appName}
-        {msg meaning="Games.pond" desc="IBID"}Pond{/msg}
+        {msg meaning="Games.pond" desc="IBID"}Pond Online{/msg}
       {/param}
     {/call}
   </h1>
